Use SupportedLanguage also from the playground.
