+++ topiary/Cargo.toml
@@ -19,6 +19,7 @@ license.workspace = true
 [dependencies]
 # For now we just load the tree-sitter language parsers statically.
 # Eventually we will want to dynamically load them, like Helix does.
+clap = { workspace = true, features = ["derive"] }
 itertools = { workspace = true }
 log = { workspace = true }
 pretty_assertions = { workspace = true }
